requirements.txt
fastapi>=0.116,<1.0
uvicorn>=0.35,<1.0
requests>=2.32,<3.0
Pillow>=11.0,<13.0
python-multipart>=0.0.20,<1.0
httpx>=0.28,<1.0
cryptography>=45,<46
beautifulsoup4>=4.14,<5.0
readability-lxml>=0.8.4,<1.0
